[back:dev]: add context in handlers,service,repo

+++ auth/internal/config/database.go
@@ -2,7 +2,7 @@ package config
 
 type DatabaseConfig struct {
 	Host string `yaml:"host"`
-	Port int    `yaml:"port"`
+	Port uint16 `yaml:"port"`
 
 	Username string `yaml:"username"`
 	Password string `yaml:"password"`
